Split Query provider into controller/provider
